Add files via upload
Very simple program created to keep track of steps I've taken in an OSCP environment. This is version 1 and will be modified over time. Goal will be to have a program that you can run, it'll ask you what you've done, and lay out your order for you for easy to use documentation.